refactor PrivateDirectories for overlap #17577
ensure language file reverted
