+++ plugins/Diagnostics/lang/en.json
@@ -22,7 +22,6 @@
         "BrowserTriggeredArchivingEnabled": "For optimal performance and a speedy Matomo, it is highly recommended to set up a crontab to automatically archive your reports, and to disable browser triggering in the Matomo settings. %1$sLearn more.%2$s",
         "NoDataForReportArchivingNotRun": "The archiving of your reports hasn't been executed recently, %1$slearn more about how to generate your reports.%2$s",
         "RecommendedPrivateDirectories": "Recommended Private Directories",
-        "RecommendPrivateFiles": "We recommend that some files and directories are private.",
         "RequiredPrivateDirectories": "Required Private Directories",
         "PrivateDirectoryManualCheck": "Please open the URLs manually in a browser to see if you can access it. If you can, you might need to modify your server configuration as these files/directories should not be accessible via a browser from the Internet or Intranet.",
         "PrivateDirectoryInternetDisabled": "We couldn't check if the following URLs are accessible because internet features are disabled on this Matomo.",
